+++ src/SelectFilesV2.tsx
@@ -1,13 +1,43 @@
-// SelectFiles.tsx
+// Filename: SelectFiles.tsx
+
 import React, { useEffect, useState } from "react";
-import { Box, Typography } from "@mui/material";
+import { Typography } from "@mui/material";
 import { AgGridReact } from "ag-grid-react";
 import { ColDef } from "ag-grid-community";
+import clsx from "clsx";
+import { makeStyles } from "@mui/styles";
 import "ag-grid-community/styles/ag-grid.css";
 import "ag-grid-community/styles/ag-theme-alpine.css";
 import { FileRowData, OrderType, SummaryType } from "./types";
 import { useLocation } from "react-router-dom";
 import OrderInfoPanel from "./OrderInfoPanel";
+import GlobalHeader from "./GlobalHeader"; // Assuming you have a GlobalHeader component
+
+const useStyles = makeStyles(() => ({
+  routeRoot: {
+    display: "flex",
+    flex: "10 auto",
+    backgroundColor: "#FAFAFA",
+    width: "100%",
+    height: "auto",
+  },
+  loader: {
+    display: "flex",
+    justifyContent: "center",
+    alignItems: "center",
+    height: "100vh", // Full viewport height
+    width: "100vw", // Full viewport width
+    position: "fixed", // Optional: Use if you want the loader to be fixed on the screen
+  },
+  mainContent: {
+    flexGrow: 1,
+    padding: 16, // Assuming theme.spacing(2) equals 16px
+  },
+  agGridContainer: {
+    height: "100%",
+    width: "100%",
+  },
+}));
 
 const defaultSummaryInfo: SummaryType = {
   calculatedGross: "N/A",
@@ -24,6 +54,7 @@ const SelectFiles: React.FC = () => {
   const orderData = location.state as OrderType | null;
   const [rowData, setRowData] = useState<FileRowData[]>([]);
   const [isLeftPanelVisible, setIsLeftPanelVisible] = useState(true);
+  const classes = useStyles();
 
   const columnDefs: ColDef[] = [
     { headerName: "File Number", field: "fileNumber", checkboxSelection: true },
@@ -65,46 +96,40 @@ const SelectFiles: React.FC = () => {
   }
 
   return (
-    <Box
-      sx={{
-        display: "flex",
-        height: "calc(100vh - 64px)",
-        marginTop: "64px",
-      }}
-    >
-      <OrderInfoPanel
-        isVisible={isLeftPanelVisible}
-        onCollapse={() => setIsLeftPanelVisible(false)}
-        onExpand={() => setIsLeftPanelVisible(true)}
-        orderInfo={orderData}
-        summaryInfo={defaultSummaryInfo}
-      />
-      <Box sx={{ flexGrow: 1, padding: 2 }}>
-        {rowData.length === 0 ? (
-          <Typography
-            variant="body1"
-            color="textSecondary"
-            align="center"
-            style={{ padding: "20px" }}
-          >
-            No Files to Display
-          </Typography>
-        ) : (
-          <Box
-            className="ag-theme-alpine"
-            style={{ height: "100%", width: "100%" }}
-          >
-            <AgGridReact
-              columnDefs={columnDefs}
-              rowData={rowData}
-              rowSelection="multiple"
-              animateRows={true}
-              pagination={true}
-            />
-          </Box>
-        )}
-      </Box>
-    </Box>
+    <>
+      <div className={clsx(classes.routeRoot, "routeRoot")}>
+        <GlobalHeader />
+        <OrderInfoPanel
+          isVisible={isLeftPanelVisible}
+          onCollapse={() => setIsLeftPanelVisible(false)}
+          onExpand={() => setIsLeftPanelVisible(true)}
+          orderInfo={orderData}
+          summaryInfo={defaultSummaryInfo}
+        />
+        <div className={classes.mainContent}>
+          {rowData.length === 0 ? (
+            <Typography
+              variant="body1"
+              color="textSecondary"
+              align="center"
+              style={{ padding: "20px" }}
+            >
+              No Files to Display
+            </Typography>
+          ) : (
+            <div className={clsx("ag-theme-alpine", classes.agGridContainer)}>
+              <AgGridReact
+                columnDefs={columnDefs}
+                rowData={rowData}
+                rowSelection="multiple"
+                animateRows={true}
+                pagination={true}
+              />
+            </div>
+          )}
+        </div>
+      </div>
+    </>
   );
 };
